video: fb: correct the fb size parse
- Remove the type cast for frame buffer size otherwise
  frame buffer size overflows and truncates to small value.

Bug 200424274

Reviewed-on: https://git-master.nvidia.com/r/1772156
Tested-by: Suresh Venkatachalam <skathirampat@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/video/tegra/dc/of_dc.c
@@ -2808,8 +2808,8 @@ struct tegra_dc_platform_data *of_dc_parse_platform_data(
 	}
 
 	if (!of_property_read_u32(np, "nvidia,fbmem-size", &temp)) {
-		pdata->fb->fbmem_size = (int)temp;
-		OF_DC_LOG("fbmem size %d\n", pdata->fb->fbmem_size);
+		pdata->fb->fbmem_size = temp;
+		OF_DC_LOG("fbmem size %u\n", pdata->fb->fbmem_size);
 	}
 
 	if (!of_property_read_u32(np, "nvidia,fb-flags", &temp)) {
